fix(start): don't run task immediately if using default framerate
When a task is scheduled at 60fps, it will already run by default on the first iteration of the loop
after scheduling it, which occurs in the same tick as the scheduling. If we run it as soon as its
scheduled, we end up running the task twice in a single tick.